requirements.txt
absl-py==0.11.0
astunparse==1.6.3
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
filelock==3.0.12
Flask==1.1.2
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.33.2
gunicorn==20.0.4
h5py==2.10.0
idna==2.10
importlib-metadata==3.1.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
Keras-Preprocessing==1.1.2
Markdown==3.3.3
MarkupSafe==1.1.1
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.7
pandas==1.1.4
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
regex==2020.11.13
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
sacremoses==0.0.43
sentencepiece==0.1.91
six==1.15.0
tensorboard-plugin-wit==1.7.0
tensorflow-cpu==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
tokenizers==0.9.3
tqdm==4.54.0
transformers==3.5.1
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
